+++ src/layout/worker/layout.worker.ts
@@ -4,6 +4,22 @@
  */
 import Layout from '../index';
 import { LAYOUT_MESSAGE } from './layoutConst';
+import { World } from '@antv/g-webgpu';
+
+const gpuVariableMapping = {
+  k: 'u_K',
+  k2: 'u_K2',
+  gravity: 'u_Gravity',
+  clusterGravity: 'u_ClusterGravity',
+  speed: 'u_Speed',
+  maxDisplace: 'u_MaxDisplace',
+  clustering: 'u_Clustering',
+  attributeArray: 'u_AttributeArray',
+  centerX: 'u_CenterX',
+  centerY: 'u_CenterY',
+  maxEdgePerVetex: 'MAX_EDGE_PER_VERTEX',
+  nodeNum: 'VERTEX_COUNT'
+}
 
 interface Event {
   type: string;
@@ -14,7 +30,7 @@ const ctx: Worker = self as any;
 
 function isLayoutMessage(event: Event) {
   const { type } = event.data;
-  return type === LAYOUT_MESSAGE.RUN;
+  return type === LAYOUT_MESSAGE.RUN || type === LAYOUT_MESSAGE.GPURUN;
 }
 
 function handleLayoutMessage(event: Event) {
@@ -35,6 +51,41 @@ function handleLayoutMessage(event: Event) {
       layoutMethod.execute();
       ctx.postMessage({ type: LAYOUT_MESSAGE.END, nodes });
       layoutMethod.destroy();
+      break;
+    }
+
+    case LAYOUT_MESSAGE.GPURUN: {
+      const { builtData, attributeArray, canvas, layoutCfg = {} } = event.data;
+
+      const {
+        maxIteration,
+        nodeNum,
+        gCode
+      } = layoutCfg;
+
+
+      const world = new World({
+        engineOptions: {
+          supportCompute: true,
+        }
+      });
+
+      const compute = world.createComputePipeline({
+        shader: gCode,
+        dispatch: [nodeNum, 1, 1],
+        maxIteration,
+        onCompleted: (finalParticleData) => {
+          // 传递数据给主线程
+          ctx.postMessage({ type: LAYOUT_MESSAGE.END, finalParticleData });
+          world.destroy();
+        },
+      });
+
+      world.setBinding(compute, 'u_Data', builtData);
+      world.setBinding(compute, 'u_AttributeArray', attributeArray);
+      Object.keys(layoutCfg).forEach(key => {
+        world.setBinding(compute, gpuVariableMapping[key], layoutCfg[key]);
+      });
       break;
     }
     default:
