+++ src/layout/worker/layoutConst.ts
@@ -13,4 +13,5 @@ export const LAYOUT_MESSAGE = {
   ERROR: 'LAYOUT_ERROR',
   // layout tick, used in force directed layout
   TICK: 'LAYOUT_TICK',
+  GPURUN: 'GPU_LAYOUT_RUN'
 };
